requirements.txt
beautifulsoup4==4.13.3
Flask==3.1.0
Flask_Cors==5.0.0
langchain==0.3.19
langchain_community==0.3.18
langchain_core==0.3.37
langchain_experimental==0.3.4
langchain_google_genai==2.0.9
langchain_openai==0.3.6
langchain_text_splitters==0.3.6
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.4
sentence-transformers==3.0.1
pypdf==4.3.0
faiss-cpu==1.9.0
